Add support for TFE Uthernet emulation
On the GSPort emulator, the Uthernet Ethernet card is emulated using The Final
Ethernet (TFE), which appears to have been originally written for a C64 emulator
and adapted to GSPort.  This commit ports TFE use into the MAME/MESS system
enabling Ethernet on the MESS Apple IIGS emulator.

To use:
* Compile from source
* MESS must be run as root, must be SUID root, or must have the cap_net_raw and
  cat_net_admin capabilities set (ala Wireshark)
* Run the IIGS emulator with -sl3 uthernet -uthernet_intf <interface>, where
  <interface> is an Ethernet interface name from "ifconfig"
* Configure Uthernet as usual

Note this expects slot 3 is Uthernet and you *must* use an Ethernet interface
for the underlying interface.